I create a delete retire goal function for my backend
