+++ server/src/controller/tables/retireTable/retireTable.ts
@@ -1,7 +1,7 @@
 import { env } from "custom-env";
 env(true);
 import pg from "pg";
-import { CreateRetireGoal, RetirementGoalsBackEnd, UpdateRetireGoal, UpdateRetiretTitle } from "../../controllerTypes/retireTypes.js";
+import { CreateRetireGoal, DeleteFromRetireGoal, RetirementGoalsBackEnd, UpdateRetireGoal, UpdateRetiretTitle } from "../../controllerTypes/retireTypes.js";
 import { Request, Response } from "express";
 
 
@@ -96,6 +96,26 @@ export async function update_RetireTable_Name(req:UpdateRetiretTitle, res:Respon
     res.status(200).json("You successfully updated your title :)")
 
   }catch(e){
+    console.log(e);
+    res.status(400).json({ msg: e.message });
+  }
+}
+
+export async function delete_Retire_Goal(req:DeleteFromRetireGoal, res:Response){
+  try {
+    const { type, id } = req.query;
+    const userId = req.userId;
+
+    
+
+    if (type !== "Retirement") return res.status(404).json({ msg: "Type of goal needed to delete is wrong" });
+    if (!id) return res.status(404).json({ msg: "Theres no user id provided" });
+
+    const text = "DELETE FROM retire WHERE creator = $1 AND type = $2 AND id = $3";
+    const values = [userId, type, id];
+    await pool.query(text, values);
+    res.send("Deleted goal");
+  } catch (e) {
     console.log(e);
     res.status(400).json({ msg: e.message });
   }
